over

+++ src/main/java/com/example/gameshoprestful/DTO/ItemDTO.java
@@ -11,7 +11,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class ItemDTO {
     private String name;
-    private BigDecimal price;
+    private double price;
 
     public String getName() {
         return name;
@@ -21,11 +21,11 @@ public class ItemDTO {
         this.name = name;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
